+++ ExpandWorld/ExpandWorld.cs
@@ -12,7 +12,7 @@ public class ExpandWorld : BaseUnityPlugin
 {
   public const string GUID = "expand_world";
   public const string NAME = "Expand World";
-  public const string VERSION = "1.27";
+  public const string VERSION = "1.28";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable
@@ -35,6 +35,8 @@ public class ExpandWorld : BaseUnityPlugin
       Directory.CreateDirectory(YamlDirectory);
     ConfigWrapper wrapper = new("expand_config", Config, ConfigSync);
     Configuration.Init(wrapper);
+    Logger.LogInfo(Path.GetFullPath(Configuration.BlueprintLocalFolder));
+    Logger.LogInfo(Path.GetFullPath(Configuration.BlueprintGlobalFolder));
     Harmony harmony = new(GUID);
     harmony.PatchAll();
     SetupWatcher();